+++ CHANGELOG.md
@@ -14,7 +14,11 @@
 - TT-1917 - Add print summary feature
 - TT-1965 - Content change
 - TT-2001 - Allow only one commodity to be selected
+- TT-2001 - Persist business and personal details
+- TT-2004 - Add link to find commodity code information page
+- TT-1993 - Use DIT sectors
 
 ### Bug fixes
 
 - TT-1992 - Fixed display of commodity codes in commodity step
+- TT-2020 - Show error message if save for later has expired